pyproject.toml
[project]
name = "shagr"
version = "0.1.0"
description = "Project for SHAGR LLC"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-stubs[compatible-mypy]>=5.2.0",
    "djangorestframework>=3.16.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]

[tool.mypy]
igmore_missing_imports = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
exclude = ["tests", ".venv"]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "server.settings"
